instagram-mcp/requirements.txt
fastmcp>=0.4.1
python-dotenv>=1.0.1
requests>=2.31.0